refactor: cherry picking just the Band-specific component changes
