--- packages/components/bolt-band/__examples__/band--card-collection.twig
@@ -1,2 +1,0 @@
-{% set pattern = "card" %}
-{% extends "@bolt/band--collection.twig" %}
