--- packages/components/bolt-band/__examples__/band--teaser-collection.twig
@@ -1,2 +1,0 @@
-{% set layout = "teaser" %}
-{% extends "@bolt/band--collection.twig" %}
