+++ packages/components/bolt-band/index.js
@@ -1,0 +1,5 @@
+import { polyfillLoader } from '@bolt/core/polyfills';
+
+polyfillLoader.then(res => {
+  import(/* webpackMode: 'eager', webpackChunkName: 'bolt-band' */ './src/band');
+});
