+++ packages/components/bolt-band/index.scss
@@ -1,0 +1,1 @@
+@import 'src/band.scss';
