+++ packages/components/bolt-band/package.json
@@ -1,16 +1,20 @@
 {
   "name": "@bolt/components-band",
-  "version": "2.3.0",
-  "sourceFilesFrom": "0.9.8",
   "description": "Band Component in Bolt",
-  "main": "src/band.js",
-  "style": "src/band.scss",
+  "license": "MIT",
+  "version": "2.3.0",
+  "homepage": "https://boltdesignsystem.com",
+  "repository": "https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-band",
+  "bugs": "https://github.com/bolt-design-system/bolt/issues",
+  "style": "index.scss",
+  "main": "index.js",
+  "twig": "src/band.twig",
+  "schema": "band.schema.yml",
   "keywords": [
     "bolt design system",
     "bolt",
-    "css framework",
     "design system",
-    "objects"
+    "components"
   ],
   "maintainers": [
     {
@@ -24,10 +28,6 @@
       "web": "http://mikemai.net/"
     }
   ],
-  "homepage": "https://boltdesignsystem.com",
-  "license": "MIT",
-  "repository": "https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-band",
-  "bugs": "https://github.com/bolt-design-system/bolt/issues",
   "publishConfig": {
     "access": "public"
   },
@@ -38,8 +38,7 @@
     "@bolt/components-grid": "^2.3.0"
   },
   "peerDependencies": {
-    "@bolt/components-grid": "^2.1.0-beta.0"
+    "@bolt/components-grid": "^2.3.0"
   },
-  "schema": "band.schema.yml",
   "gitHead": "3d198feea5f73c63913a956a2a077e229ad8d462"
 }
